Add a new `constr` property, to represent the constructor function the prototype came from at the time the object was constructed.
This also necessitates refactoring everything that deals with creating objects. This is too many different places to explain concisely. However, the general rule is that all object construction now requires both the prototype and it's associated constructor; a `SystemConstructors` bit has been added and is available in all the same places `SystemPrototypes` is.