Move `css::Name` and `css::Keyword` into `html`, since they specifically have to do with how HTML uses CSS.
Believe it or not, it is in fact possible to use CSS as a generic styling language outside of HTML. Notably, GTK themes are CSS stylesheets that are applied to a *decidedly* not HTML environment. The `css` module is being genericized specifically to allow it's reuse outside of this particular HTML-in-Flash context.